fix all warning add SimpleAR example

+++ windageLib/src/Algorithms/HomographyEstimator.cpp
@@ -91,7 +91,7 @@ void DecomposeHomographyToRT(CvMat *intrinsic, CvMat *Homography, CvMat *RT)
 		// Search next orthonormal matrix:
 		if( cvNorm( &m_CinvH1, NULL, CV_L2, NULL ) != 0 )
 		{
-			float lambda = 1.0f/cvNorm( &m_CinvH1, NULL, CV_L2, NULL );
+			float lambda = (float)(1.0f/cvNorm( &m_CinvH1, NULL, CV_L2, NULL ));
 
 			// Create normalized R1 & R2:
 			cvGEMM( invIntrinsic, &m_H1, lambda, NULL, 0, &m_R1, 0 );
